clean(recs): remove old recommendations route, view, serializers, model, tests
