+++ src/apps/recommendations/serializers.py
@@ -2,19 +2,7 @@ from rest_framework import serializers
 
 from apps.recipes.models import Recipe
 from apps.recipes.serializers import RecommendedRecipeSerializer
-from apps.recommendations.models import Recommendation, RecipeRecommendation
-
-
-# TODO: remove
-class RecommendationSerializer(serializers.ModelSerializer):
-    recipe_id = serializers.CharField(source='recipe.id', read_only=True)
-    recipe_title = serializers.CharField(source='recipe.title', read_only=True)
-    recipe_description = serializers.CharField(source='recipe.description', read_only=True)
-    image = serializers.ImageField(source='recipe.image', read_only=True)
-
-    class Meta:
-        model = Recommendation
-        exclude = ['id', 'profile', 'recipe']
+from apps.recommendations.models import RecipeRecommendation
 
 
 class RecipeRecommendationSerializer(serializers.ModelSerializer):
